+++ Strava_Server_v1/src/es/deusto/ingenieria/sd/auctions/server/data/dao/SessionDAO.java
@@ -6,15 +6,19 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityTransaction;
 import javax.persistence.Query;
+import javax.persistence.TypedQuery;
 
 import es.deusto.ingenieria.sd.auctions.server.data.domain.Session;
 import es.deusto.ingenieria.sd.auctions.server.data.domain.User;
+
+
 
 public class SessionDAO extends DataAccessObjectBase implements IDataAccessObject<Session> {
 
 	private static SessionDAO instance;
 
-	private SessionDAO() {}
+	private SessionDAO() {
+	}
 
 	public static SessionDAO getInstance() {
 		if (instance == null) {
@@ -25,20 +29,25 @@ public class SessionDAO extends DataAccessObjectBase implements IDataAccessObjec
 
 	@Override
 	public void store(Session object) {
-		Session storedObject = instance.find(String.valueOf(object.getTitle()));
-
+		
+		Session storedObject = null;
+		if(object.getId() != null)
+		{
+			storedObject = instance.find(String.valueOf(String.valueOf(object.getId())));
+		}
+		
 		EntityManager em = emf.createEntityManager();
 		EntityTransaction tx = em.getTransaction();
 
 		try {
 			tx.begin();
-			
+
 			if (storedObject != null) {
 				em.merge(object);
 			} else {
 				em.persist(object);			
 			}
-			
+
 			tx.commit();
 		} catch (Exception ex) {
 			System.out.println("  $ Error storing Session: " + ex.getMessage());
@@ -48,7 +57,8 @@ public class SessionDAO extends DataAccessObjectBase implements IDataAccessObjec
 			}
 
 			em.close();
-		}	}
+		}
+	}
 
 	@Override
 	public void delete(Session object) {
@@ -56,19 +66,19 @@ public class SessionDAO extends DataAccessObjectBase implements IDataAccessObjec
 		EntityTransaction tx = em.getTransaction();
 
 		try {
-            tx.begin();
+			tx.begin();
 
-            // Retrieve the managed session
-            Session managedSession = em.find(Session.class, object.getTitle());
-
-         // Remove associated user_session records
-        
-
-
-            em.remove(managedSession);
+			// Retrieve the managed session
+			System.out.println("Object Id " + object.getId());
+			Session managedSession = em.find(Session.class, String.valueOf(object.getId()));
+			// Remove associated user_session records
+			if (managedSession != null) {
+                em.remove(managedSession);
+            }
 
 			tx.commit();
 		} catch (Exception ex) {
+			
 			System.out.println("  $ Error removing an Session: " + ex.getMessage());
 			ex.printStackTrace();
 		} finally {
@@ -91,7 +101,7 @@ public class SessionDAO extends DataAccessObjectBase implements IDataAccessObjec
 
 		try {
 			tx.begin();
-			
+
 			Query q = em.createQuery("SELECT b FROM Session b");
 			session = (List<Session>) q.getResultList();
 
@@ -109,18 +119,52 @@ public class SessionDAO extends DataAccessObjectBase implements IDataAccessObjec
 		return session;
 	}
 
+	public List<Session> getSessionsForUser(User user) {
+		EntityManager em = emf.createEntityManager();
+		EntityTransaction tx = em.getTransaction();
+		List<Session> sessions = null;
+		
+		try {
+			tx.begin();
+			
+			TypedQuery<Session> query = null;
+			// Get sessions for the given user
+			query = em.createQuery("SELECT s FROM Session s WHERE s.user = :user",
+					Session.class);
+			query.setParameter("user", user);
+			sessions = query.getResultList();
+			tx.commit();
+
+		} catch (Exception ex) {
+			System.out.println("  $ Error retrieving all the Session for User: " + ex.getMessage());
+		} finally {
+			if (tx != null && tx.isActive()) {
+				tx.rollback();
+			}
+
+			em.close();
+		}
+		return sessions;
+	}
+
 	@Override
 	public Session find(String param) {
 		EntityManager em = emf.createEntityManager();
 		EntityTransaction tx = em.getTransaction();
-		
-		Session result = null; 
+
+		Session result = null;
 
 		try {
 			tx.begin();
-						
-			result = (Session) em.find(Session.class, param);
 			
+			 if (param != null) {
+		            Long sessionId = Long.parseLong(param);
+		            result = (Session) em.find(Session.class, sessionId);
+		        } else {
+		            // Handle the case where param is null (e.g., log a message or throw an exception)
+		            System.out.println("  $ Error: Session ID is null");
+		        }
+
 			tx.commit();
 		} catch (Exception ex) {
 			System.out.println("  $ Error querying an Session by Id: " + ex.getMessage());
